Separate the summarizer into its own class so we can filter which chunks get summarized
